Realloc now creates a free block when reducing size
When reducing the size of a memory block using realloc it would not create a new free block to be used by the memory allocator. Fixed an issue with free where it would not properly short circuit the loop when looking for the closest previous free block and also patched free where it would not properly iterate for the previously mentioned free block.