Added Err handling, Free mem tracking, auto defrag
In malloc added a check to make sure that the memory pool could theoretically hold the new occupied block. Added the auto-defragmentation logic to malloc which will run defragment if the freeMemory var is larger than the size value of the new occupied block. Added Error handling for defragment which will produce an error if it was called when there was no free memory and also if it was run but no free blocks were moved. Added a check for realloc which will make sure that the resized memory block can theoretically fit in the memory pool based of the freeMemory variable.