Path github workflow
Remove the seperation of run unit tests into two seperate jobs as there is no reason to send a build file to the Build with tests job since its just going to be rebuilt with tests, this can be changed later if the build process is to be seperated from the run job. Change runDepPackage to use dpkg rather than apt install since the package contains source files that need to be built first before being installed also specify that the file to be installed is a .deb file.