Patch fetching logic
Alter the code so that it doesn't use gtest_FOUND to check whether or not gtest is on the system or not, this is because gtest_FOUND is not assigned when gtest is retrieved via fetch so they have been seperated. The .deb package now requires users to libgtest already on their systems when installing the .deb package due to fetch not being viable through .deb.