Implement patches
Fix an issue with free() where it would not properly handle updating ptrs when there is a free block on both ends. Fix an issue with TestConnections() where it wouldn't begin with a free chunk if occHead is not a nullptr. Rewrite realloc() return lines to return single ptrs and not double ptrs.